+++ ic-agent/src/agent/route_provider/dynamic_routing/snapshot/latency_based_routing.rs
@@ -5,8 +5,11 @@ use std::{
 
 use rand::Rng;
 
-use crate::agent::route_provider::dynamic_routing::{
-    health_check::HealthCheckStatus, node::Node, snapshot::routing_snapshot::RoutingSnapshot,
+use crate::agent::route_provider::{
+    dynamic_routing::{
+        health_check::HealthCheckStatus, node::Node, snapshot::routing_snapshot::RoutingSnapshot,
+    },
+    RoutesStats,
 };
 
 // Determines the size of the sliding window used for storing latencies and availabilities of nodes.
@@ -157,6 +160,7 @@ fn compute_score(
 pub struct LatencyRoutingSnapshot {
     nodes_with_metrics: Vec<NodeWithMetrics>,
     existing_nodes: HashSet<Node>,
+    healthy_nodes: HashSet<Node>,
     window_weights: Vec<f64>,
     window_weights_sum: f64,
     use_availability_penalty: bool,
@@ -174,6 +178,7 @@ impl LatencyRoutingSnapshot {
         Self {
             nodes_with_metrics: vec![],
             existing_nodes: HashSet::new(),
+            healthy_nodes: HashSet::new(),
             use_availability_penalty: true,
             window_weights,
             window_weights_sum,
@@ -302,6 +307,12 @@ impl RoutingSnapshot for LatencyRoutingSnapshot {
                 self.nodes_with_metrics.len() - 1
             });
 
+        if health.is_healthy() {
+            self.healthy_nodes.insert(node.clone());
+        } else {
+            self.healthy_nodes.remove(node);
+        }
+
         self.nodes_with_metrics[idx].add_latency_measurement(health.latency());
 
         self.nodes_with_metrics[idx].score = compute_score(
@@ -314,6 +325,10 @@ impl RoutingSnapshot for LatencyRoutingSnapshot {
 
         true
     }
+
+    fn routes_stats(&self) -> RoutesStats {
+        RoutesStats::new(self.existing_nodes.len(), Some(self.healthy_nodes.len()))
+    }
 }
 
 #[cfg(test)]
@@ -323,15 +338,18 @@ mod tests {
         time::Duration,
     };
 
-    use crate::agent::route_provider::dynamic_routing::{
-        health_check::HealthCheckStatus,
-        node::Node,
-        snapshot::{
-            latency_based_routing::{
-                compute_score, weighted_sample, LatencyRoutingSnapshot, NodeWithMetrics,
+    use crate::agent::route_provider::{
+        dynamic_routing::{
+            health_check::HealthCheckStatus,
+            node::Node,
+            snapshot::{
+                latency_based_routing::{
+                    compute_score, weighted_sample, LatencyRoutingSnapshot, NodeWithMetrics,
+                },
+                routing_snapshot::RoutingSnapshot,
             },
-            routing_snapshot::RoutingSnapshot,
         },
+        RoutesStats,
     };
 
     #[test]
@@ -344,6 +362,7 @@ mod tests {
         assert!(!snapshot.has_nodes());
         assert!(snapshot.next_node().is_none());
         assert!(snapshot.next_n_nodes(1).is_empty());
+        assert_eq!(snapshot.routes_stats(), RoutesStats::new(0, Some(0)));
     }
 
     #[test]
@@ -359,6 +378,7 @@ mod tests {
         assert!(snapshot.nodes_with_metrics.is_empty());
         assert!(!snapshot.has_nodes());
         assert!(snapshot.next_node().is_none());
+        assert_eq!(snapshot.routes_stats(), RoutesStats::new(0, Some(0)));
     }
 
     #[test]
@@ -370,6 +390,7 @@ mod tests {
         let node = Node::new("api1.com").unwrap();
         let health = HealthCheckStatus::new(Some(Duration::from_secs(1)));
         snapshot.existing_nodes.insert(node.clone());
+        assert_eq!(snapshot.routes_stats(), RoutesStats::new(1, Some(0)));
         // Check first update
         let is_updated = snapshot.update_node(&node, health);
         assert!(is_updated);
@@ -377,6 +398,7 @@ mod tests {
         let node_with_metrics = snapshot.nodes_with_metrics.first().unwrap();
         assert_eq!(node_with_metrics.score, (2.0 / 1.0) / 2.0);
         assert_eq!(snapshot.next_node().unwrap(), node);
+        assert_eq!(snapshot.routes_stats(), RoutesStats::new(1, Some(1)));
         // Check second update
         let health = HealthCheckStatus::new(Some(Duration::from_secs(2)));
         let is_updated = snapshot.update_node(&node, health);
@@ -399,6 +421,7 @@ mod tests {
         assert_eq!(snapshot.nodes_with_metrics.len(), 1);
         assert_eq!(snapshot.existing_nodes.len(), 1);
         assert!(snapshot.next_node().is_none());
+        assert_eq!(snapshot.routes_stats(), RoutesStats::new(1, Some(0)));
     }
 
     #[test]