+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added `CanisterInfo` to `MgmtMethod`.
 
+* Extended `RouteProvider` trait with `fn routes_stats()`, returning the number of total and healthy routes.
+
 ## [0.39.2] - 2024-12-20
 
 * Bumped `ic-certification` to `3.0.0`.